+++ src/main/java/raisetech/StudentManagement/Controller/StudentController.java
@@ -1,0 +1,33 @@
+package raisetech.StudentManagement.Controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import raisetech.StudentManagement.data.Student;
+import raisetech.StudentManagement.data.StudentCourses;
+import raisetech.StudentManagement.service.StudentService;
+
+import java.util.List;
+
+@RestController
+public class StudentController {
+
+    private StudentService service;
+
+    @Autowired
+    public StudentController(StudentService service) {
+        this.service = service;
+    }
+
+
+    @GetMapping("/studentList")
+    public List<Student> getStudentList() {
+        return service.searchStudentList();
+    }
+
+    @GetMapping("/studentCourseList")
+    public List<StudentCourses> getStudentCourseList() {
+        return service.searchStudentCourseList();
+    }
+
+}
